Move individual components to their own modules
